requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.0
emoji==0.5.4
lxml==4.5.0
pymongo==3.10.1
PyYAML==5.3.1
requests==2.23.0
retrying==1.3.3
wxpy==0.3.9.8
